[setup-usecase] Move Problem classes to the base/result package
