+++ app/src/main/java/br/com/sommelier/data/repository/AuthRepositoryImpl.kt
@@ -4,19 +4,18 @@ import arrow.core.Either
 import arrow.core.left
 import arrow.core.right
 import br.com.sommelier.domain.repository.AuthRepository
-import br.com.sommelier.util.AlreadySignedOutUserProblem
-import br.com.sommelier.util.DeleteUserProblem
-import br.com.sommelier.util.NullResultProblem
-import br.com.sommelier.util.NullUserProblem
-import br.com.sommelier.util.Problem
-import br.com.sommelier.util.ReauthenticateUserProblem
-import br.com.sommelier.util.RegisterUserProblem
-import br.com.sommelier.util.SendEmailVerificationProblem
-import br.com.sommelier.util.SendPasswordResetEmailProblem
-import br.com.sommelier.util.SignInUserProblem
-import br.com.sommelier.util.UpdateUserEmailProblem
-import br.com.sommelier.util.UpdateUserPasswordProblem
-import com.google.firebase.auth.AuthCredential
+import br.com.sommelier.base.result.AlreadySignedOutUserProblem
+import br.com.sommelier.base.result.DeleteUserProblem
+import br.com.sommelier.base.result.NullResultProblem
+import br.com.sommelier.base.result.NullUserProblem
+import br.com.sommelier.base.result.Problem
+import br.com.sommelier.base.result.ReauthenticateUserProblem
+import br.com.sommelier.base.result.RegisterUserProblem
+import br.com.sommelier.base.result.SendEmailVerificationProblem
+import br.com.sommelier.base.result.SendPasswordResetEmailProblem
+import br.com.sommelier.base.result.SignInUserProblem
+import br.com.sommelier.base.result.UpdateUserEmailProblem
+import br.com.sommelier.base.result.UpdateUserPasswordProblem
 import com.google.firebase.auth.EmailAuthProvider
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
