+++ app/src/main/java/br/com/sommelier/data/repository/UserRepositoryImpl.kt
@@ -8,13 +8,13 @@ import br.com.sommelier.domain.mapper.toData
 import br.com.sommelier.domain.mapper.toDomain
 import br.com.sommelier.domain.model.UserDomain
 import br.com.sommelier.domain.repository.UserRepository
-import br.com.sommelier.util.AddDocumentProblem
-import br.com.sommelier.util.DeleteDocumentProblem
+import br.com.sommelier.base.result.AddDocumentProblem
+import br.com.sommelier.base.result.DeleteDocumentProblem
 import br.com.sommelier.util.FirestoreCollections.UID_FIELD
-import br.com.sommelier.util.GetDocumentProblem
-import br.com.sommelier.util.NotFoundDocumentProblem
-import br.com.sommelier.util.Problem
-import br.com.sommelier.util.UpdateDocumentProblem
+import br.com.sommelier.base.result.GetDocumentProblem
+import br.com.sommelier.base.result.NotFoundDocumentProblem
+import br.com.sommelier.base.result.Problem
+import br.com.sommelier.base.result.UpdateDocumentProblem
 import com.google.firebase.firestore.CollectionReference
 import kotlinx.coroutines.tasks.await
