+++ app/src/main/java/br/com/sommelier/domain/repository/UserRepository.kt
@@ -2,7 +2,7 @@ package br.com.sommelier.domain.repository
 
 import arrow.core.Either
 import br.com.sommelier.domain.model.UserDomain
-import br.com.sommelier.util.Problem
+import br.com.sommelier.base.result.Problem
 
 interface UserRepository {
     suspend fun saveUser(user: UserDomain): Either<Problem, Unit>
