+++ app/src/test/java/br/com/sommelier/data/AuthRepositoryImplTest.kt
@@ -2,17 +2,17 @@ package br.com.sommelier.data
 
 import android.text.TextUtils
 import br.com.sommelier.data.repository.AuthRepositoryImpl
-import br.com.sommelier.util.AlreadySignedOutUserProblem
-import br.com.sommelier.util.DeleteUserProblem
-import br.com.sommelier.util.NullResultProblem
-import br.com.sommelier.util.NullUserProblem
-import br.com.sommelier.util.ReauthenticateUserProblem
-import br.com.sommelier.util.RegisterUserProblem
-import br.com.sommelier.util.SendEmailVerificationProblem
-import br.com.sommelier.util.SendPasswordResetEmailProblem
-import br.com.sommelier.util.SignInUserProblem
-import br.com.sommelier.util.UpdateUserEmailProblem
-import br.com.sommelier.util.UpdateUserPasswordProblem
+import br.com.sommelier.base.result.AlreadySignedOutUserProblem
+import br.com.sommelier.base.result.DeleteUserProblem
+import br.com.sommelier.base.result.NullResultProblem
+import br.com.sommelier.base.result.NullUserProblem
+import br.com.sommelier.base.result.ReauthenticateUserProblem
+import br.com.sommelier.base.result.RegisterUserProblem
+import br.com.sommelier.base.result.SendEmailVerificationProblem
+import br.com.sommelier.base.result.SendPasswordResetEmailProblem
+import br.com.sommelier.base.result.SignInUserProblem
+import br.com.sommelier.base.result.UpdateUserEmailProblem
+import br.com.sommelier.base.result.UpdateUserPasswordProblem
 import com.google.android.gms.tasks.Task
 import com.google.firebase.auth.AuthResult
 import com.google.firebase.auth.EmailAuthCredential
